+++ __tests__/__src__/ITestPropertiesSchema.ts
@@ -73,6 +73,7 @@ export interface ITestPropertiesSchema {
          * a PDS member containing IEFBR14 JCL that your user ID can submit
          */
         iefbr14Member: string,
+        sleepMember: string;
     };
     uss: {
         /**