+++ __tests__/__system__/cli/submit/data-set/__scripts__/command/command_submit_data_set_wait.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+dsn=$1
+option=$2
+wait=$3
+set -e
+if [ "$2" == "--wait" ] 
+then
+bright zos-ftp submit data-set "$dsn" "$option" "$wait"
+else 
+ bright zos-ftp submit data-set "$dsn" "$option"   
+fi