+++ __tests__/__system__/cli/submit/data-set/cli.submit.data-set.test.ts
@@ -14,8 +14,8 @@ import { FTPConfig } from "../../../../../src/api/FTPConfig";
 import { TestEnvironment } from "../../../../__src__/environment/TestEnvironment";
 import { runCliScript } from "../../../../__src__/TestUtils";
 import * as path from "path";
+import { isNull } from "util";
 
-let dsname: string;
 let user: string;
 let connection: any;
 
@@ -51,14 +51,72 @@ describe("submit job from data set command", () => {
 
     it("should be able to submit a job from a local file and see the job name and job id", async () => {
 
-        // download the appropriate JCL content from the data set
         const iefbr14DataSet = testEnvironment.systemTestProperties.jobs.iefbr14Member;
-        const jclFilePath = testEnvironment.workingDir + "/iefbr14.txt";
         const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set.sh", testEnvironment, [iefbr14DataSet]);
         expect(result.stderr.toString()).toEqual("");
         expect(result.status).toEqual(0);
         expect(result.stdout.toString()).toContain("jobid");
         expect(result.stdout.toString()).toContain("jobname");
+    });
+
+    it("should be able to submit a job from a local file with wait option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const option = "--wait";
+        const wait = "3,5";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set_wait.sh", testEnvironment, [sleepDataSet,option,wait]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Waiting for job completion.");
+        expect(result.output.toString()).toContain("rc:");
+        expect(result.output.toString()).toContain("finished");
+    });
+
+    it("should give a syntax error if the wait value is invalid", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const option = "--wait";
+        const wait = "3,";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set_wait.sh", testEnvironment, [sleepDataSet,option,wait]);
+        expect(result.output.toString()).toContain("comma-separated numeric values");
+        expect(result.output.toString()).toContain("Syntax error:");
+        expect(result.status).toEqual(0);
+    });
+
+    it("should be able to submit a job from a local file but not finished within specified wait option", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const option = "--wait";
+        const wait = "1,2";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set_wait.sh", testEnvironment, [sleepDataSet,option,wait]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Submitted job successfully");
+        expect(result.output.toString()).toContain("Job is still running.");
+        expect(result.output.toString()).toContain("Please using the following command to check its status later:");
+    });
+
+    it("should be able to submit a job from a local file with wait-for-output option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const option = "--wait-for-output";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set_wait.sh", testEnvironment, [sleepDataSet,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Waiting for job completion.");
+        expect(result.output.toString()).toContain("rc:");
+        expect(result.output.toString()).toContain("finished");
+    });
+
+    it("should be able to submit a job from a local file with wait-for-active option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const option = "--wait-for-active";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_data_set_wait.sh", testEnvironment, [sleepDataSet,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).not.toContain("rc:");
+        expect(result.output.toString()).toContain("ACTIVE");
     });
 
     it("should give a syntax error if the data set to submit is omitted", async () => {