+++ __tests__/__system__/cli/submit/local-file/__scripts__/command/command_submit_local_file_wait.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+localFile=$1
+option=$2
+wait=$3
+set -e
+
+if [ "$2" == "--wait" ] 
+then
+bright zos-ftp submit local-file "$localFile" "$option" "$wait"
+else
+bright zos-ftp submit local-file "$localFile" "$option"
+fi