+++ __tests__/__system__/cli/submit/stdin/cli.submit.stdin.test.ts
@@ -49,7 +49,6 @@ describe("submit job from stdin command", () => {
 
     it("should be able to submit a job from standard in and see the job name and job id", async () => {
 
-        // download the appropriate JCL content from the data set
         const iefbr14DataSet = testEnvironment.systemTestProperties.jobs.iefbr14Member;
         const iefbr14Content = (await connection.getDataset(iefbr14DataSet)).toString();
         const jclFilePath = testEnvironment.workingDir + "/iefbr14.txt";
@@ -61,4 +60,78 @@ describe("submit job from stdin command", () => {
         expect(result.stdout.toString()).toContain("jobname");
     });
 
+    it("should be able to submit a job from a local file with wait option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option ="--wait";
+        const wait = "3,5";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,option,wait]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Waiting for job completion.");
+        expect(result.output.toString()).toContain("rc:");
+    });
+
+    it("should give a syntax error if the wait value is invalid", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option ="--wait";
+        const wait = "3,";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,option,wait]);
+        expect(result.output.toString()).toContain("comma-separated numeric values");
+        expect(result.output.toString()).toContain("Syntax error:");
+        expect(result.status).toEqual(0);
+    });
+
+    it("should be able to submit a job from a local file but not finished within specified wait option", async () => {
+
+        // download the appropriate JCL content from the data set
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option ="--wait";
+        const wait = "1,2";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,option,wait]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Submitted job successfully");
+        expect(result.output.toString()).toContain("Job is still running.");
+        expect(result.output.toString()).toContain("Please using the following command to check its status later:");
+    });
+
+    it("should be able to submit a job from a local file with wait-for-output option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option = "--wait-for-output";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).toContain("Waiting for job completion.");
+        expect(result.output.toString()).toContain("rc:");
+        expect(result.output.toString()).toContain("finished");
+    });
+
+    it("should be able to submit a job from a local file with wait-for-active option and get rc successfully", async () => {
+
+        const sleepDataSet = testEnvironment.systemTestProperties.jobs.sleepMember;
+        const sleepContent = (await connection.getDataset(sleepDataSet)).toString();
+        const jclFilePath = testEnvironment.workingDir + "/sleep.txt";
+        await IO.writeFileAsync(jclFilePath, sleepContent);
+        const option = "--wait-for-active";
+        const result = runCliScript(__dirname + "/__scripts__/command/command_submit_stdin_wait.sh", testEnvironment, [jclFilePath,option]);
+        expect(result.stderr.toString()).toEqual("");
+        expect(result.status).toEqual(0);
+        expect(result.output.toString()).not.toContain("rc:");
+        expect(result.output.toString()).toContain("ACTIVE");
+    });
 });