+++ src/cli/submit/data-set/DataSet.Handler.ts
@@ -9,28 +9,15 @@
  *
  */
 
-import { IO } from "@zowe/imperative";
-import { JobUtils } from "../../../api/JobUtils";
-import { FTPBaseHandler } from "../../../FTPBase.Handler";
 import { IFTPHandlerParams } from "../../../IFTPHandlerParams";
+import { SubmitJobHandler } from "../SubmitJobHandler";
 
-export default class SubmitJobFromLocalFileHandler extends FTPBaseHandler {
+export default class SubmitJobFromLocalFileHandler extends SubmitJobHandler {
     public async processFTP(params: IFTPHandlerParams): Promise<void> {
-        let jobDetails: any;
-        let jobid: string;
         this.log.debug("Submitting a job from data set '%s'. Downloading before submitting...", params.arguments.dataSet);
         const dsContent = (await params.connection.getDataset("'" + params.arguments.dataSet + "'")).toString();
         this.log.debug("Downloaded data set '%s'. Submitting...", params.arguments.dataSet);
-
-        jobid = await params.connection.submitJCL(dsContent);
-        jobDetails = await JobUtils.findJobByID(jobid, params.connection);
-        this.log.info("Submitted job successfully, jobname(jobid): %s(%s)", jobDetails.jobname, jobDetails.jobid);
-        params.response.data.setObj(jobDetails);
-        params.response.format.output({
-            output: jobDetails,
-            format: "object",
-            fields: ["jobid", "jobname", "owner", "status"]
-        });
+        return this.submitJCL(dsContent, params);
     }
 }