+++ src/cli/submit/local-file/LocalFile.Handler.ts
@@ -10,25 +10,15 @@
  */
 
 import { IO } from "@zowe/imperative";
-import { JobUtils } from "../../../api/JobUtils";
-import { FTPBaseHandler } from "../../../FTPBase.Handler";
 import { IFTPHandlerParams } from "../../../IFTPHandlerParams";
+import { SubmitJobHandler } from "../SubmitJobHandler";
 
-export default class SubmitJobFromLocalFileHandler extends FTPBaseHandler {
+
+export default class SubmitJobFromLocalFileHandler extends SubmitJobHandler {
     public async processFTP(params: IFTPHandlerParams): Promise<void> {
-        let jobDetails: any;
-        let jobid: string;
         this.log.debug("Submitting a job from local file %s", params.arguments.file);
         const fileContent = IO.readFileSync(params.arguments.file);
-        jobid = await params.connection.submitJCL(fileContent);
-        jobDetails = await JobUtils.findJobByID(jobid, params.connection);
-        this.log.info("Submitted job successfully, jobname(jobid): %s(%s)", jobDetails.jobname, jobDetails.jobid);
-        params.response.data.setObj(jobDetails);
-        params.response.format.output({
-            output: jobDetails,
-            format: "object",
-            fields: ["jobid", "jobname", "owner", "status"]
-        });
+        return this.submitJCL(fileContent.toString(), params);
     }
 }