+++ src/cli/submit/local-file/LocalFile.definition.ts
@@ -27,6 +27,10 @@ export const SubmitLocalFileDefinition: ICommandDefinition = {
             description: "Submit a job from the local file \"my_build_jcl.txt\" and print only the job ID",
             options: "\"my_build_jcl.txt\" --rff jobid --rft string"
         },
+        {
+            description: "Submit a job from the local file \"my_build_jcl.txt\" and wait for job complete.",
+            options: "\"my_build_jcl.txt\" --wait 5,12"
+        },
     ],
     positionals: [{
         name: "file",
@@ -34,7 +38,32 @@ export const SubmitLocalFileDefinition: ICommandDefinition = {
         type: "existingLocalFile",
         required: true
     }],
-    options: [],
+    options: [
+        {
+            name: "wait",
+            aliases: ["w"],
+            description: "Specify job query interval and max times of querying job status. " +
+            "The format of this option is comma-separated numeric values. For example, '5,12' means queries job status every 5 seconds for 12 times at most.",
+            type: "string",
+            required: false,
+            conflictsWith: ["wait-for-active", "wait-for-output"]
+        },
+        {
+            name: "wait-for-output",
+            aliases: ["wfo"],
+            description: "Wait for the job to enter OUTPUT status before completing the command.",
+            type: "boolean",
+            required: false,
+            conflictsWith: ["wait-for-active", "wait"]
+        },
+        {
+            name: "wait-for-active",
+            aliases: ["wfa"],
+            description: "Wait for the job to enter ACTIVE status before completing the command.",
+            type: "boolean",
+            required: false,
+            conflictsWith: ["wait-for-output", "wait"]
+        }],
     profile:
         {optional: ["zftp"]},
     outputFormatOptions: true